+++ docs_web/lib/dialogs/dialog.dart
@@ -52,14 +52,15 @@ showDialog(
               active: _dialogActive,
               onPressed: () async {
                 setState(() => _dialogActive = true);
-                await showDialog(
-                  context: context,
-                  barrierDismissible: false,
+                late DialogController dialogController;
+                dialogController = showDialog(
+                  context,
+                  dismissible: false,
                   builder: (context) => Dialog(
-                    menus: [
+                    actions: [
                       DialogAction(
                         title: 'Close',
-                        onPressed: () => Dialog.close(context),
+                        onPressed: () => dialogController.close(),
                       ),
                     ],
                     title: Text('Lorem Ipsum'),
@@ -77,6 +78,8 @@ Eos sed fuga neque. Rerum adipisci ducimus et aspernatur in. Atque sequi fugiat 
                   ),
                 );
 
+                final _ = await dialogController.closed;
+
                 setState(() => _dialogActive = false);
               },
             ),
@@ -93,15 +96,16 @@ Eos sed fuga neque. Rerum adipisci ducimus et aspernatur in. Atque sequi fugiat 
               onPressed: () async {
                 setState(() => _dialogDismissActive = true);
 
-                await showDialog(
-                  context: context,
-                  barrierDismissible: true,
+                final dialog = showDialog(
+                  context,
                   builder: (context) => Dialog(
                     title: Text('Lorem Ipsum'),
                     body: Text(
                         'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.'),
                   ),
                 );
+
+                final _ = await dialog.closed;
 
                 setState(() => _dialogDismissActive = false);
               },